+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Next
 
 
 * Added ``myst_parser`` parsers using the ``myst-parser`` library.
+* Added ``docutils_rst`` parsers using the ``docutils`` library.
 * Drop support for Python 3.10.
 
 2026.02.27